+++ core/dao/comm_dao.go
@@ -40,7 +40,7 @@ func (c *CommonDao) GetQrTemplates() []*CommQrTemplate {
 }
 
 // 获取二维码模板
-func (c *CommonDao) GetQrTemplate(id int64) *CommQrTemplate {
+func (c *CommonDao) GetQrTemplate(id int32) *CommQrTemplate {
 	for _, v := range c.GetQrTemplates() {
 		if v.Id == id {
 			return v
@@ -60,7 +60,7 @@ func (c *CommonDao) SaveQrTemplate(q *CommQrTemplate) error {
 	if q.BgImage == "" {
 		return errors.New("二维码背景图片为空")
 	}
-	_, err := orm.Save(dOrm, q, q.Id)
+	_, err := orm.Save(dOrm, q, int(q.Id))
 	if err == nil {
 		dSto.Del(qrStoKey)
 	}
@@ -68,7 +68,7 @@ func (c *CommonDao) SaveQrTemplate(q *CommQrTemplate) error {
 }
 
 // 删除二维码模板
-func (c *CommonDao) DelQrTemplate(id int64) error {
+func (c *CommonDao) DelQrTemplate(id int32) error {
 	err := dOrm.DeleteByPk(CommQrTemplate{}, id)
 	if err == nil {
 		dSto.Del(qrStoKey)
